requirements.txt
vosk>=0.3.45
piper-tts
pyserial>=3.5
sounddevice>=0.4.6
PyYAML>=6.0
numpy>=1.21
pymorphy3>=1.0
rapidfuzz>=3.0
navec>=0.10.0
pyVoIP>=1.6.0
